TA-30 Improve readme
Card TA-30 is about making a deployment to the SAML mock, whilst this is
an important part of the shared mock IDP instance, the main use of this
is for developers in other teams to set themselves up with the mock with
minimum effort from the get access team.  These readme changes will help
clarify what and how the mock IDP can be used.